wip auth

+++ packages/api/src/convex/auth/usecases/logout.usecase.ts
@@ -1,4 +1,3 @@
-import type { EmptyObject } from '@game/shared';
 import { MutationUseCase } from '../../usecase';
 
 export type LogoutInput = never;
@@ -10,7 +9,10 @@ export class LogoutUseCase extends MutationUseCase<LogoutInput, LogoutOutput> {
   static INJECTION_KEY = 'logoutUseCase' as const;
 
   async execute(): Promise<LogoutOutput> {
-    await this.ctx.sessionRepo.delete(this.ctx.session._id);
+    if (this.ctx.session) {
+      await this.ctx.sessionRepo.delete(this.ctx.session._id);
+    }
+
     return { success: true };
   }
 }
